refactor: moved assets-loading into assets_loader.py module
This way we will always keep our assets inside memory. It may be bad,
in some rare cases (optimisation shenanigans). But its also good for
the most. But basically, what we achieved there is removal of duplicate
data loading, that could happen if multiple characters of same type
has been initialized to exist on screen together