feat: initial animations support
For now its not perfect, but it works - for each spritesheet, we now
may have proper animations. The way its handled is not te best thing
in the world, nor the most efficient. Also all animations will play
with same speed, which may be a flaw, depending on future needs. But,
for now, I will keep related functionality "as is"