feat: player receive damage on collision and can die
Also refactored entity 2d objects from dictionary into classes. Idk
if this is a good idea, but I wanted to do that for quite a long time
anyway. #TODO: add invincibility frames, coz right now player keep
getting damage for each frame it collides with enemy. Which is, to say
lightly, not something we need