feat: projectiles without billboard, projectile resize
First means projectiles now render in parallel to the floor, which
should fix the issue with projectile shooting downwards be partially
covered with floor tile. And second allows to set literally any size
of projectile node. In future I will need to add similar thing for
creatures, but it will require playing with offsets (coz of billboard),
and right now thats not my top priority