feat: added music player module
This is but nasty workaround on top of panda's default music manager,
but it should solve the issue with "clicky" noise on tracks switch, by
introducing fade out/fade in effects and ability to crossfade music.
Plus it potentially introduced ability to play multiple music tracks
at once, but I didnt test it. In future, I may add some things on top
of it, such as ability to loop just certain parts of music tracks.
But for now it will do "as is"