feat: initial support for toml-based enemy skills
This is extremely WIP and many features are broken right now. But base
is there - enemy is now able to cast skills configured from tomls and
player will only get damage on collision with enemy projectiles and not
with enemies themselves.
But ye - projectiles chasing after caster, auto-resize on non-standard
caster size, fixes for animations - all of these are #TODO