feat: internal entity2d node hierarchy
Because sometimes we may want to attach some elements to node (text,
shadow, head, etc), it made sense to stop referring to inner node of
animatedobject and create new one, the following is attached to. And
also I've added optional self.visuals, enabled with "visuals_node" bool
of entity2d, which adds separate node for visuals, serving as child of
main mode I've mentioned above, but parent of all visuals attached to
it. This way, it now should be possible to affect all visuals together,
while also not affecting things that shouldnt (for example, only change
direction of visuals, while keeping floating text the same). And, cuz
projectile dont need that fancy stuff, its disabled for it.

Or at least that was the idea. Hopefully things will work as intended