feat: animated heads support
With update to p3dss, its now possible to reset sprites to default if
specified amount of time has passed. Which allowed for support of
animated heads, even with single-sprite actions (what it means is head
will change its animation together with body, if it has more than one
animation of active "hairstyle" attached to it. Then, after a while, go
back to default state)