refactor: split interface.py into package
Now there are separate files for custom ui parts and interfaces that
may (or may not) utilize them. Parts may eventually end up in separate
panda3d library, idk. The whole point of this change is to debloat
original interface.py, which started to become one of the largest
modules out there. Further changes to be expected in future