refactor: FlatMap improvements
- Fixed issue with floor tiles blinking on movement by using different
(blurry) texture filter and enabling mipmapping. This kills the idea of
having pixel-perfect floor, but it kinda seems to be general appoach in
gameindustry, based on multiple 2.5D games I've checked
- Slightly reworked FlatMap's structure: it now accepts ability to
attach map to different scene (default=render) and only generates map
after generate(), not on init
- Flat map now slightly adjust its size to make floor tiles scale
perfectly (the bigger is the tile - the bigger the difference with
requested size will be). I dont see this as issue, since there are
plans to rework this whole mechanism anyway
- Enemies now spawn not outside of map's edges, but on map's edges
- Updated required p3dss version 0.3.0 -> 0.3.1, to fix issue with
entities randomly getting black borders for no reason